fix: Update nonce input to show queued txs and reset button
